Moved motd and login so they are similar to emote_d.o
The idea behind this is to move them out of the way so that
normal users of the mudlib will not have to do fancy tricks with git because
they modified the motd or the login message etc....

Note if any of you are keeping track of things, when you do a git pull
this will remove the files for you so you'll want to copy them back
into place.  See the INSTALLING doc for details.  I added a warning
if the files are missing for those two.