Big patch by Aidil,
- move user privileges out of secure_d and into user_d
- move domains out of secure_d into a new domain_d
- whole lot of small changes to adapt for those 2 things

On top of that, it adds 4 new afuns:
- spit_array()
- set_property()
- property()
- remove_property()

Documentation for split_array is included, will write something for the
property functions. besides the obvious things, they allow for 'secure'
properties, which is kinda neat. They are used in their secure variation to
mark users for 'auto admin' and 'ALL_USERS_WIZ' which will be picked up once
the player is fully created.

Besides this, I extended the 'rebuild' command and included an updated version
of the m_getopt library I sent you earlier.

Also included are a few fixes:

- ensure privilege changes/domain membership propagate to the player object as
well as user and ftp sessions for online players
- ftp sessions now run with regular user privileges (this also makes it work
for domain members)
- fixed a number of runtime errors in ftp_d
- ftp_d now gives feedback when it can't do things instead of acting as if
everything works, but not doing anything.
- disable the broken include tracking in compiler_d as it was a potential
cause of runtime errors
- fix a whole bunch of 'out of ticks' errors during the loading of objects
- stop filter_array() from complaining when a filter function doesn't return
an int