add another set_this_player
if an object in a players inventory has a heart_beat, this_player is
set to the player holding the object.  Instances where this applies are
called with a 0-second call_out so that this_player is not retained on
the same thread dispatching events (or this_player would be stored
where it shouldn't).