Major reorganization of afun/afun-game and safun
Simplified by consolidating afun-game into afuns and safuns.
Got rid of afun-game.
Added a safun lib file to sys/lib/.  This takes care of loading any
defines, variables, and declaring functions in the right order.
Made adjustments to docs to reflect changes.
I put as much as possible inside of safun for the sake of builders to
have access to a good reference source.  Anything that dealt with
security went into afun.
Made change to compiler_d and driver to reflect new safun lib file, and
the new paths to some safuns that used to be afuns.
I think that sums it up.