Revert "add quest data rehashing"
This reverts commit 15801259cc296da61fbd02900bc080233f326050.

Conflicts:
	lib/daemons/quest_d.c
	lib/sys/cmds/admin/questadm.c

This was trying to use two different methods of handling quests and was
very confusing.  I want it to be simple for people checking out the mudlib.
This method seems to mesh with the rest of the lib currently as well, so
I'm reverting it.  If we decide we want to go with a text file that people
can edit instead, it fully needs to be converted to that method not try
to do both.  I think a config file is really more complicated though,
much easier to just use the command line to add a quest.

One key thing is I've tried to make the mudlib in a way that all "user"
data will not interfere with what's in git.  An example of this is the
emote's emote_d.o is stuck in contrib, this way people can use it and
we have a copy in git, but if people add emote's for their mudlib it
doesn't cause a conflict in git that they have to work around.

I feel making yet another datafile for quests makes this more complicated
as well, that was part of the decision to revert this change.