Added afun new_object from Aidil
This is needed to fix a number of bugs when using lightweight objects (LWOs).
Its implementation is a slight variation on the clone_object() afun, below is
what I use in my own gurbalib derived kernel and should work for gurbalib as
well. Put it in /kernel/lib/afun/new_object.c and #include that file from
/kernel/lib/auto.c

Bugs I noticed that made me add this are console_msg() not working in LWOs and
create() not being called in LWOs on creation (note that destructing() will
never be called in LWOs as they are not destructed explicitly)

From the looks of the code, and from what I remember from how the auto object
works, there should be more bugs in LWO handling that will be fixed by adding
this afun.